% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{length,APV-method}
\alias{length,APV-method}
\title{Amalgams Pseudo-Vector Length}
\usage{
\S4method{length}{APV}(x)
}
\arguments{
\item{x}{an instance of \code{APV}}
}
\value{
the number of amalgams (permutations with replacement) in pseudo-vector \code{x}
}
\description{
Get the length of an \code{APV} instance.
}
\details{
Since \code{x} contains all the \code{k}-amalgams of objects in vector
\code{items}, \code{length(x)} will return \code{length(items) ^ k)}.
}
\seealso{
Permutations Pseudo-Vector \code{\link{ppv}}

Combinations Pseudo-Vector \code{\link{cpv}}

Selections Pseudo-Vector \code{\link{spv}}

Subsets Pseudo-Vector \code{\link{sspv}}
}

