% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{length,CPV-method}
\alias{length}
\alias{length,CPV-method}
\title{Combinations Pseudo-Vector Length}
\usage{
\S4method{length}{CPV}(x)
}
\arguments{
\item{x}{an instance of \code{CPV}}
}
\value{
the number of combinations in pseudo-vector \code{x}
}
\description{
Get the length of a \code{CPV} instance.
}
\details{
Since \code{x} contains all the \code{k}-combinations of objects in vector
\code{items}, \code{length(x)} will return \code{choose(length(items), k)}.
}
\seealso{
Combinations Pseudo-Vector \code{\link{cpv}}
}

