% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{length,PPV-method}
\alias{length,PPV-method}
\title{Permutations Pseudo-Vector Length}
\usage{
\S4method{length}{PPV}(x)
}
\arguments{
\item{x}{an instance of \code{PPV}}
}
\value{
the number of permutations in pseudo-vector \code{x}
}
\description{
Get the length of a \code{PPV} instance.
}
\details{
Since \code{x} contains all the \code{k}-permutations of objects in vector
\code{items}, \code{length(x)} will return
\code{choose(length(items), k) * factorial(k)}.
}
\seealso{
Combinations Pseudo-Vector \code{\link{cpv}}

Amalgams Pseudo-Vector \code{\link{apv}}

Selections Pseudo-Vector \code{\link{spv}}

Subsets Pseudo-Vector \code{\link{sspv}}
}

