% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{length,SSPV-method}
\alias{length,SSPV-method}
\title{Subsets Pseudo-Vector Length}
\usage{
\S4method{length}{SSPV}(x)
}
\arguments{
\item{x}{an instance of \code{SSPV}}
}
\value{
the number of subsets in pseudo-vector \code{x}
}
\description{
Get the length of an \code{SSPV} instance.
}
\details{
Since \code{x} contains all the subsets of objects in vector
\code{items}, \code{length(x)} will return \code{2 ^ length(items)}.
}
\seealso{
Permutations Pseudo-Vector \code{\link{ppv}}

Combinations Pseudo-Vector \code{\link{cpv}}

Amalgams Pseudo-Vector \code{\link{apv}}

Selections Pseudo-Vector \code{\link{spv}}
}

