% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,APV-method}
\alias{[,APV-method}
\title{Retrieve an Amalgam by Index}
\usage{
\S4method{[}{APV}(x, i, j, drop)
}
\arguments{
\item{x}{an instance of \code{APV}.}

\item{i}{an index specifying the position of the sought amalgam}

\item{j}{not used.}

\item{drop}{not used.}
}
\value{
the amalgam located at position \code{i} in pseudo-vector \code{x}
}
\description{
Access an amalgam (permutation with replacement) stored in an \code{APV} instance by index.
}
\details{
The amalgam at index \code{i} of pseudo-vector \code{x} is not actually
stored in memory but calculated as needed. The extract method is used solely
for interpretation.
}
\seealso{
Permutations Pseudo-Vector \code{\link{ppv}}

Combinations Pseudo-Vector \code{\link{cpv}}

Selections Pseudo-Vector \code{\link{spv}}

Subsets Pseudo-Vector \code{\link{sspv}}
}

