% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,CPV-method}
\alias{[,CPV-method}
\title{Retrieve a Combination by Index}
\usage{
\S4method{[}{CPV}(x, i, j, drop)
}
\arguments{
\item{x}{an instance of \code{CPV}.}

\item{i}{an index specifying the position of the sought combination.}

\item{j}{not used.}

\item{drop}{not used.}
}
\value{
the combination located at position \code{i} in pseudo-vector \code{x}
}
\description{
Access a combination stored in a \code{CPV} instance by index.
}
\details{
The combination at index \code{i} of pseudo-vector \code{x} is not actually
stored in memory but calculated as needed. The extract method is used solely
for interpretation.
}
\seealso{
Permutations Pseudo-Vector \code{\link{ppv}}

Amalgams Pseudo-Vector \code{\link{apv}}

Selections Pseudo-Vector \code{\link{spv}}

Subsets Pseudo-Vector \code{\link{sspv}}
}

