% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,PPV-method}
\alias{[,PPV-method}
\title{Retrieve a Permutation by Index}
\usage{
\S4method{[}{PPV}(x, i, j, drop)
}
\arguments{
\item{x}{an instance of \code{PPV}.}

\item{i}{an index specifying the position of the sought permutation.}

\item{j}{not used.}

\item{drop}{not used.}
}
\value{
the permutation located at position \code{i} in pseudo-vector \code{x}
}
\description{
Access a permutation stored in a \code{PPV} instance by index.
}
\details{
The permutation at index \code{i} of pseudo-vector \code{x} is not actually
stored in memory but calculated as needed. The extract method is used solely
for interpretation.
}
\seealso{
Combinations Pseudo-Vector \code{\link{cpv}}

Amalgams Pseudo-Vector \code{\link{apv}}

Selections Pseudo-Vector \code{\link{spv}}

Subsets Pseudo-Vector \code{\link{sspv}}
}

