% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,SPV-method}
\alias{[,SPV-method}
\title{Retrieve a Selection by Index}
\usage{
\S4method{[}{SPV}(x, i, j, drop)
}
\arguments{
\item{x}{an instance of \code{SPV}.}

\item{i}{an index specifying the position of the sought selection.}

\item{j}{not used.}

\item{drop}{not used.}
}
\value{
the selection located at position \code{i} in pseudo-vector \code{x}
}
\description{
Access a selection (combination with replacement) stored in an \code{SPV} instance by index.
}
\details{
The selection at index \code{i} of pseudo-vector \code{x} is not actually
stored in memory but calculated as needed. The extract method is used solely
for interpretation.
}
\seealso{
Permutations Pseudo-Vector \code{\link{ppv}}

Combinations Pseudo-Vector \code{\link{cpv}}

Amalgams Pseudo-Vector \code{\link{apv}}

Subsets Pseudo-Vector \code{\link{sspv}}
}

