% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{[,SSPV-method}
\alias{[,SSPV-method}
\title{Retrieve a Subset by Index}
\usage{
\S4method{[}{SSPV}(x, i, j, drop)
}
\arguments{
\item{x}{an instance of \code{SSPV}.}

\item{i}{an index specifying the position of the sought amalgam}

\item{j}{not used.}

\item{drop}{not used.}
}
\value{
the subset located at position \code{i} in pseudo-vector \code{x}
}
\description{
Access asubset stored in an \code{SSPV} instance by index.
}
\details{
The subset at index \code{i} of pseudo-vector \code{x} is not actually
stored in memory but calculated as needed. The extract method is used solely
for interpretation.
}
\seealso{
Permutations Pseudo-Vector \code{\link{ppv}}

Combinations Pseudo-Vector \code{\link{cpv}}

Amalgams Pseudo-Vector \code{\link{apv}}

Selections Pseudo-Vector \code{\link{spv}}
}

